Update all dependencies to latest compatible versions
Updated dependency versions across all workspace packages while respecting
version constraints from DEPENDENCIES.md:

Root package.json updates:
- TypeScript: ~5.4.3 → ^5.9.3 (latest 5.x)
- Playwright: ^1.42.1 → ^1.56.1
- Webpack: ^5.91.0 → ^5.102.1
- Prettier: ^3.2.5 → ^3.6.2
- Rollup: ^4.14.0 → ^4.52.5
- Various @types/* packages updated to latest
- Many other dev dependencies updated to latest minor/patch versions

modules/core/package.json updates:
- @colyseus/schema: ^2.0.30 → ^2.0.37 (stayed in 2.0.x)
- colyseus.js: ^0.15.18 → ^0.15.28 (stayed in 0.15.x)
- detect-collisions: ^9.5.3 → ^9.27.6 (stayed in 9.x)
- xstate: ^5.9.1 → ^5.24.0
- ws: ^8.16.0 → ^8.18.3
- tslib: ^2.6.2 → ^2.8.1

modules/browser/package.json updates:
- pixi.js: ^7.1.2 → ^7.4.3 (stayed in 7.x)
- @pixi/graphics-extras: ^7.1.2 → ^7.4.3
- @xstate/react: ^4.1.0 → ^4.1.3
- react-dropzone: ^14.2.3 → ^14.3.8
- hotkeys-js: ^3.13.7 → ^3.13.15

modules/server/package.json updates:
- colyseus: ^0.15.15 → ^0.15.57 (stayed in 0.15.x)
- @colyseus/monitor: ^0.15.6 → ^0.15.8
- @colyseus/ws-transport: ^0.15.1 → ^0.15.3
- express: ^4.19.2 → ^4.21.2

Preserved version pins (as per DEPENDENCIES.md):
- golden-layout: 1.5.9 (v2.x incompatible)
- @arwes/*: 1.0.0-alpha.19 (alpha pinned)
- react/react-dom: ^17.0.2 (v18+ needs testing)
- colyseus ecosystem: 0.15.x (0.16.x has breaking changes)
- pixi.js: 7.x (8.x has breaking changes)

Fixed TypeScript 5.9 compatibility issue:
- Added explicit Uint8Array type assertions in Buffer conversions for
  node:zlib functions (game-state-serialization.ts)

All tests passing. Build successful.

+++ modules/server/src/serialization/game-state-serialization.ts
@@ -7,7 +7,7 @@ const do_gzip = promisify(gzip);
 const do_unzip = promisify(unzip);
 
 export async function schemaToString(fragment: Schema) {
-    const zipped = await do_gzip(Buffer.from(fragment.encodeAll()));
+    const zipped = await do_gzip(Buffer.from(fragment.encodeAll()) as Uint8Array);
     return zipped.toString('base64');
 }
 type Constructor<T extends Schema> = new (...args: never[]) => T;
@@ -17,12 +17,12 @@ export async function stringToSchema<T extends Schema>(ctor: Constructor<T>, ser
 }
 
 export async function stringToSchemaObject<T extends Schema>(obj: T, serialized: string) {
-    const unzipped = await do_unzip(Buffer.from(serialized, 'base64'));
+    const unzipped = await do_unzip(Buffer.from(serialized, 'base64') as Uint8Array);
     obj.decode([...unzipped]);
     return obj;
 }
 
 export async function getUnzipped(zipped: string) {
-    const unzipped = await do_unzip(Buffer.from(zipped, 'base64'));
+    const unzipped = await do_unzip(Buffer.from(zipped, 'base64') as Uint8Array);
     return unzipped.toString('base64');
 }
